requirements.txt
Flask>0.9
pickle
sklearn
requests
numpy
json
azure.storage.blob
sklearn.neighbors